backend/requirements.txt
django
djangorestframework
django-cors-headers
django-ratelimit
spacy
textblob
tiktoken
openai
celery
redis
django-redis
jsonschema
language-tool-python
validators
requests
numpy
scikit-learn
python-dotenv
gunicorn
social-auth-app-django
razorpay
resend
psycopg2-binary
whitenoise
